refactor: extract MarkdownWithMath component and replace SolidMarkdown usages
Introduces a shared MarkdownWithMath component that bundles SolidMarkdown
with KaTeX auto-rendering (via CDN auto-render.mjs) and the KaTeX CSS link,
eliminating duplicated inline KaTeX setup across arena and page components.
