+++ src/routes/arenas/codegolf.tsx
@@ -1,4 +1,3 @@
-import {Link} from '@solidjs/meta';
 import {useSearchParams} from '@solidjs/router';
 import {Alert, Box, Button, ButtonGroup, Card, CardContent, CircularProgress, Container, FormControl, Grid, InputLabel, Link as LinkUi, MenuItem, Paper, Select, Stack, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, TextField, Typography} from '@suid/material';
 import {SelectChangeEvent} from '@suid/material/Select';
@@ -6,15 +5,14 @@ import {blue, green, red} from '@suid/material/colors';
 import {stripIndent} from 'common-tags';
 import dayjs from 'dayjs';
 import {addDoc, collection, CollectionReference, doc, DocumentReference, getFirestore, orderBy, query, serverTimestamp, where} from 'firebase/firestore';
-import remarkGfm from 'remark-gfm';
 import {floor, zip} from 'remeda';
 import {useFirebaseApp, useFirestore} from 'solid-firebase';
 import {createEffect, createMemo, createSignal, For, Match, onCleanup, Show, Switch} from 'solid-js';
-import {SolidMarkdown} from 'solid-markdown';
 import {setArenaTitle, useUser} from '../arenas';
 import styles from './reversing-diff.module.css';
 import Collection from '~/components/Collection';
 import Doc from '~/components/Doc';
+import MarkdownWithMath from '~/components/MarkdownWithMath';
 import Username from '~/components/Username';
 import PageNotFoundError from '~/lib/PageNotFoundError';
 import {codegolfLanguageAllowList} from '~/lib/const';
@@ -60,25 +58,11 @@ const MainTab = (props: MainTabProps) => {
 	const [selectedAnyLanguage, setSelectedAnyLanguage] = createSignal<string | null>(null);
 	const [languageInfos, setLanguageInfos] = createSignal<{shortestSize: number | null}[]>([]);
 
-	let descriptionEl!: HTMLElement;
-
 	const executionLanguage = createMemo(() => {
 		if (selectedLanguage() !== 'anything') {
 			return selectedLanguage();
 		}
 		return selectedAnyLanguage();
-	});
-
-	createEffect(async () => {
-		// @ts-expect-error: URL import
-		// eslint-disable-next-line import/no-unresolved
-		const {default: renderMathInElement} = await import('https://cdn.jsdelivr.net/npm/katex@0.16.10/dist/contrib/auto-render.mjs');
-		renderMathInElement(descriptionEl, {
-			delimiters: [
-				{left: '$$', right: '$$', display: true},
-				{left: '$', right: '$', display: false},
-			],
-		});
 	});
 
 	const handleClickSubmit = async () => {
@@ -175,14 +159,8 @@ const MainTab = (props: MainTabProps) => {
 
 				return (
 					<>
-						<Link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/katex@0.16.10/dist/katex.min.css"/>
-						<Typography variant="body1" ref={descriptionEl}>
-							<SolidMarkdown
-								class="markdown"
-								children={config.description}
-								remarkPlugins={[remarkGfm]}
-								linkTarget="_blank"
-							/>
+						<Typography variant="body1">
+							<MarkdownWithMath content={config.description}/>
 						</Typography>
 						<Grid container spacing={2} class={styles.rule}>
 							<For each={config.testcases}>
